requirements.txt
djangorestframework
Pillow
django
